combine sys and eos
